refactor: изменить названия переменных и функций.
