+++ script/FormValidation.js
@@ -1,14 +1,3 @@
-
-const formValidationConfig = {
-  formSelector: '.popup__form',
-  inputSelector: '.popup__profile-edit',
-  submitButtonSelector: '.popup__btn',
-  inactiveButtonClass: 'popup__btn_inactive',
-  inputErrorClass: 'popup__profile-edit_type_error',
-  errorClass: 'popup__input-error_active'
-
-};
-
 class FormValidation{
   constructor(formValidationConfig, formElement){
     this._formElement = formElement;
@@ -82,9 +71,8 @@ class FormValidation{
   enableValidation() {
     this._setEventListener();
   }
-
 }
 
-export {FormValidation, formValidationConfig}
+export {FormValidation}
