+++ script/formValidationConfig.js
@@ -1,0 +1,10 @@
+const formValidationConfig = {
+  formSelector: '.popup__form',
+  inputSelector: '.popup__profile-edit',
+  submitButtonSelector: '.popup__btn',
+  inactiveButtonClass: 'popup__btn_inactive',
+  inputErrorClass: 'popup__profile-edit_type_error',
+  errorClass: 'popup__input-error_active'
+};
+
+export {formValidationConfig};
